src/train_model/requirements.txt
torch
torchvision
tqdm
kaggle
wandb
python-dotenv
scikit-learn
pandas
numpy<2
google-cloud-storage
google-cloud-aiplatform
